認可処理追加

+++ app/Http/Controllers/ArticleController.php
@@ -76,7 +76,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
-        //
+        $this->authorize($article);
         $data = ['article' => $article];
         return view('articles.edit', $data);
     }
@@ -90,7 +90,7 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
-        //
+        $this->authorize($article);
         $this->validate($request, [
             'title' => 'required|max:255',
             'body' => 'required'
@@ -109,7 +109,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
-        //
+        $this->authorize($article);
         $article->delete();
         return redirect(route('articles.index'));
     }
